Add a function that will fetch the form data
Using an event listener I first prevented the default behavior
of submiting a form. This form is only used to collect data
and work with it in the front-end. After that I assigned the
value of each of the form fields into a variable and called
the Book constructor with those variables. I then added the
book to the library array and called displayBooks.