Enable the deletion of a book
In displayBooks() I declared an index variable that will keep
track of each of the books index in the myLibrary array. I
then declared a new property bookIndex and for each book I
assigned the index to bookIndex. I declared another function
that will remove a book based on it's index and update the array
and the display.